+++ src/gemseo_pymoo/post/scatter_pareto.py
@@ -154,7 +154,7 @@ class ScatterPareto(BasePost[ScatterParetoPostSettings]):
             plot.add(pareto.f_anti_utopia, label=nadir_label, **self.prop_extra)
 
         # Plot points of interest.
-        if settings.points is not None and len(settings.points) > 0:
+        if settings.points is not None:
             if isinstance(settings.points_labels, str):
                 plot.add(
                     settings.points, label=settings.points_labels, **self.prop_interest
@@ -165,7 +165,7 @@ class ScatterPareto(BasePost[ScatterParetoPostSettings]):
         plot.do()
 
         # Create arrows.
-        if settings.plot_arrow and settings.points is not None:
+        if settings.plot_arrow:
             for point in settings.points:
                 # Arrow vector.
                 vect = point - pareto.f_utopia